Moved nearestLine_Coro tp lineGroup
I think this is a better location for the method. it would be nice if I
didn't have to do a deep copy.
